+++ evm/abi/src/abi.rs
@@ -1,48 +1,90 @@
 //! Solidity ABI abstraction.
-use crate::Input;
+
+use crate::Arg;
+use core::{convert::Infallible, str::FromStr};
+
+#[cfg(not(feature = "std"))]
+use crate::std::{String, ToString, Vec};
 
 /// Solidity ABI abstraction.
-#[derive(Debug, Clone)]
+#[derive(Clone, Debug, Default)]
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub struct Abi {
     /// ABI name.
     pub name: String,
     /// ABI type.
+    #[cfg_attr(feature = "serde", serde(rename = "type"))]
     pub ty: Type,
-    /// ABI inputs.
-    pub inputs: Vec<Input>,
+    /// An array of arguments.
+    pub inputs: Vec<Arg>,
+    /// An array of arguments, similar to inputs.
+    pub outputs: Vec<Arg>,
 }
 
 #[cfg(feature = "syn")]
 impl From<&syn::Signature> for Abi {
     fn from(sig: &syn::Signature) -> Self {
-        let args = sig.inputs.iter().filter_map(|arg| {
-            if let syn::FnArg::Typed(syn::PatType { ty, .. }) = arg {
-                Some(Input {
-                    name: sig.ident.to_string(),
-                    ty: crate::Param::from(ty),
-                })
-            } else {
-                None
-            }
-        });
+        let inputs = sig
+            .inputs
+            .iter()
+            .filter_map(|arg| {
+                if let syn::FnArg::Typed(syn::PatType { ty, .. }) = arg {
+                    Some(Arg {
+                        name: sig.ident.to_string(),
+                        ty: crate::Param::from(ty),
+                    })
+                } else {
+                    None
+                }
+            })
+            .collect();
 
+        let outputs = if let syn::ReturnType::Type(_, ty) = &sig.output {
+            vec![Arg {
+                name: sig.ident.to_string(),
+                ty: crate::Param::from(ty),
+            }]
+        } else {
+            vec![]
+        };
+
+        let name = sig.ident.to_string();
         Abi {
-            name: sig.ident.to_string(),
-            inputs: args.collect(),
-            ty: Type::Function,
+            ty: Type::from(name.as_str()),
+            name,
+            inputs,
+            outputs,
         }
     }
 }
 
 /// Solidity ABI type.
-#[derive(Debug, Clone)]
+#[derive(Clone, Debug, Default)]
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
+#[cfg_attr(feature = "serde", serde(rename_all = "lowercase"))]
 pub enum Type {
     /// Constructor ABI.
     Constructor,
     /// Function ABI.
+    #[default]
     Function,
+}
+
+impl From<&str> for Type {
+    fn from(s: &str) -> Self {
+        match s {
+            "constructor" => Type::Constructor,
+            _ => Type::Function,
+        }
+    }
+}
+
+impl FromStr for Type {
+    type Err = Infallible;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        Ok(Self::from(s))
+    }
 }
 
 impl AsRef<str> for Type {