+++ tests/if.rs
@@ -1,6 +1,6 @@
 //! if-else tests for the zink compiler.
 use anyhow::Result;
-use zinkc_filetests::Test;
+use filetests::Test;
 use zint::{Bytes32, Contract, InstructionResult};
 
 #[test]