+++ tests/loop.rs
@@ -1,7 +1,7 @@
 //! loop tests
 
 use anyhow::Result;
-use zinkc_filetests::Test;
+use filetests::Test;
 use zint::{Bytes32, Contract, InstructionResult};
 
 #[test]