+++ tests/recursion.rs
@@ -1,5 +1,5 @@
 use anyhow::Result;
-use zinkc_filetests::Test;
+use filetests::Test;
 use zint::{Bytes32, Contract};
 
 #[test]